Integration Tests: download aqua-dht builtin (#1182)
