+++ .circleci/config.yml
@@ -35,19 +35,33 @@ jobs:
             -u "$CIRCLE_API_TOKEN:")
 
             BINARY=$(echo "$ARTIFACTS" | jq -r ".items | .[] | select(.path == \"target/release/particle-node\") | \"\(.url)\"")
-            
-            curl --fail \
+            SHA256=$(sha256sum ./target/release/particle-node | awk '{ print $1 }')
+
+            OUTPUT_FILE=$(mktemp)
+            HTTP_CODE=$(curl \
+              --silent \
+              --output $OUTPUT_FILE \
+              --write-out "%{http_code}" \
               -u folex:${GITHUB_PERSONAL_TOKEN} \
               -X POST \
               -H "Accept: application/vnd.github.v3+json" \
               https://api.github.com/repos/fluencelabs/node-distro/actions/workflows/container.yml/dispatches \
               -d '{
-                "ref":"main",
+                "ref":"container_branch",
                 "inputs": {
                   "container_tag":"'${BRANCH}'",
-                  "fluence_binary":"'${BINARY}'"
+                  "version": "'${CIRCLE_BRANCH}_${CIRCLE_SHA1}'",
+                  "url": "'${BINARY}'",
+                  "sha256": "'${SHA256}'"
                 }
               }'
+            )
+            if [[ ${HTTP_CODE} -lt 200 || ${HTTP_CODE} -gt 299 ]] ; then
+              >&2 cat $OUTPUT_FILE
+              exit 22
+            fi
+            cat $OUTPUT_FILE
+            rm $OUTPUT_FILE
       - save_cache:
           paths:
             - ~/.cargo
