+++ particle-node/tests/builtins/tests/src/aqua_dht/aqua_dht.rs
@@ -14,36 +14,33 @@
  * limitations under the License.
  */
 
-use crate::aqua_dht::Record;
-use crate::load_script;
+use super::Record;
+use crate::{load_script, SERVICES};
 use connected_client::ConnectedClient;
 use created_swarm::make_swarms_with_builtins;
 
 use eyre::WrapErr;
 use maplit::hashmap;
 use serde_json::json;
-use std::path::Path;
 
 #[test]
-fn find_subscribers() {
-    let init_subscribe_script = load_script("pubsub.initTopicAndSubscribe.air");
-    let find_subscribers_script = load_script("pubsub.findSubscribers.air");
+fn put_get() {
+    let register_put_script = load_script("dht-api.registerKeyPutValue.air");
+    let get_values_script = load_script("dht-api.getValues.air");
 
-    let swarms = make_swarms_with_builtins(3, Path::new("../deploy/builtins"), None);
+    let swarms = make_swarms_with_builtins(3, SERVICES.as_ref(), None);
 
     let mut client = ConnectedClient::connect_to(swarms[0].multiaddr.clone())
         .wrap_err("connect client")
         .unwrap();
 
-    log::info!("gonna send particle");
-    // initTopicAndSubscribe(node_id: PeerId, topic: string, value: string, relay_id: ?PeerId, service_id: ?string):
-    let topic = "topic";
-    client.send_particle_ext(
-        init_subscribe_script,
+    let key = "key";
+    let put_particle = client.send_particle_ext(
+        register_put_script,
         hashmap! {
             "-relay-" => json!(client.node.to_string()),
             "node_id" => json!(client.node.to_string()),
-            "topic" => json!(topic),
+            "key" => json!(key),
             "value" => json!("value"),
             // relay_id & service_id are of Option type, so they are represented as arrays
             "relay_id" => json!([client.node.to_string()]),
@@ -51,23 +48,27 @@ fn find_subscribers() {
         },
         true,
     );
+    client
+        .wait_particle_args(&put_particle)
+        .expect("execute registerKeyPutValue");
 
-    // func findSubscribers(node_id: PeerId, topic: string) -> []Record:
-    let find_subscribers_particle = client.send_particle_ext(
-        find_subscribers_script,
+    let get_particle = client.send_particle_ext(
+        get_values_script,
         hashmap! {
             "-relay-" => json!(client.node.to_string()),
             "node_id" => json!(client.node.to_string()),
-            "topic" => json!(topic),
+            "relay_id" => json!(client.node.to_string()),
+            "key" => json!(key),
         },
         true,
     );
     let mut result = client
-        .wait_particle_args(&find_subscribers_particle)
-        .expect("execute findSubscribers");
+        .wait_particle_args(&get_particle)
+        .expect("execute getValues");
     let records = result.get_mut(0).unwrap().take();
     let records: Vec<Record> = serde_json::from_value(records).unwrap();
-    assert_eq!(records.len(), 1, "expected 1 record, got {}", records.len());
+
+    assert!(records.len() > 0);
 
     for record in records {
         assert_eq!(&record.value, "value");
